[*] разбил replacers.h на несколько файлов
