+++ sources/im4_im5/CMakeLists.txt
@@ -6,6 +6,9 @@ set(SOURCE_FILES
     complexoyaz.cpp
     labels_counter.cpp
     variables_counter.cpp
+    functional_replacers.cpp
+    free_label_replacer.cpp
+    free_var_replacer.cpp
 )
 
 add_executable(im4_im5 ${SOURCE_FILES} ${BACKWARD_ENABLE})
