+++ sources/im4_im5/functional_replacers.cpp
@@ -1,0 +1,15 @@
+#include "functional_replacers.h"
+
+namespace cyaz {
+
+FunctionalReplacer::FunctionalReplacer(ReplaceFunction function) :
+    function(std::move(function))
+{
+}
+
+std::string FunctionalReplacer::resolve(const trm::PatternStringInfo &patternStringInfo)
+{
+    return function(patternStringInfo);
+}
+
+}
