+++ sources/im4_im5/functional_replacers.h
@@ -1,0 +1,21 @@
+#pragma once
+
+#include <functional>
+#include <string>
+#include <translation_module/replacers.h>
+
+namespace cyaz {
+
+using ReplaceFunction = std::function<std::string(const trm::PatternStringInfo &patternStringInfo)>;
+
+class FunctionalReplacer : public trm::Replacer {
+public:
+    FunctionalReplacer(ReplaceFunction function);
+
+    std::string resolve(const trm::PatternStringInfo &patternStringInfo) override;
+
+private:
+    ReplaceFunction function;
+};
+
+}
