--- sources/im4_im5/replacers.cpp
@@ -1,1 +1,0 @@
-#include "replacers.h"
